+++ platform/msm_shared/dev_tree.c
@@ -58,7 +58,7 @@ extern uint32_t target_dev_tree_mem(void *fdt, uint32_t memory_node_offset);
 /* TODO: This function needs to be moved to target layer to check violations
  * against all the other regions as well.
  */
-extern int check_aboot_addr_range_overlap(uint32_t start, uint32_t size);
+extern int check_aboot_addr_range_overlap(uintptr_t start, uint32_t size);
 
 int fdt_check_header_ext(const void *fdt)
 {
@@ -484,7 +484,7 @@ void *dev_tree_appended(void *kernel, uint32_t kernel_size, void *tags)
 			break;
 		dtb_size = fdt_totalsize(&dtb_hdr);
 
-		if (check_aboot_addr_range_overlap((uint32_t)tags, dtb_size)) {
+		if (check_aboot_addr_range_overlap((uintptr_t)tags, dtb_size)) {
 			dprintf(CRITICAL, "Tags addresses overlap with aboot addresses.\n");
 			return NULL;
 		}
@@ -1289,7 +1289,7 @@ int update_device_tree(void *fdt, const char *cmdline,
 		return ret;
 	}
 
-	if (check_aboot_addr_range_overlap((uint32_t)fdt,
+	if (check_aboot_addr_range_overlap((uintptr_t)fdt,
 				(fdt_totalsize(fdt) + DTB_PAD_SIZE))) {
 		dprintf(CRITICAL, "Error: Fdt addresses overlap with aboot addresses.\n");
 		return ret;
